cambio de ruta para la lista de productos, edicion, creacion y detalle
