+++ app.js
@@ -2,6 +2,7 @@ const express = require('express');
 const path = require('path');
 
 const mainRouter = require('./routers/index');
+const productsRouter = require('./routers/products');
 
 const app = express();
 
@@ -11,6 +12,7 @@ app.set('view engine', 'ejs');
 app.set('views', path.join(__dirname, './views'));
 
 app.use('/', mainRouter);
+app.use('/products', productsRouter);
 
 app.listen(3030, () => {
     console.log('Servidor iniciado en http://localhost:3030');
