+++ models/products.json
@@ -9,7 +9,8 @@
         "img1": "papel-glace-fluo.png",
         "img2": "papel-glace-metalizado.png",
         "img3": "papel-glace-lustre.png",
-        "descripcion": null
+        "descripcion": null,
+        "categoria": "schoolSupplies"
     },
     {
         "id": 2,
@@ -21,7 +22,8 @@
         "img1": "mochila-unicornio-1.jpg",
         "img2": "mochila-unicornio-2.jpg",
         "img3": null,
-        "descripcion": null
+        "descripcion": null,
+        "categoria": "schoolSupplies"
     },
     {
         "id": 3,
@@ -33,7 +35,8 @@
         "img1": "mochila-videojuego-1.jpg",
         "img2": "mochila-videojuego-2.jpg",
         "img3": null,
-        "descripcion": null
+        "descripcion": null,
+        "categoria": "schoolSupplies"
     },
     {
         "id": 4,
@@ -45,7 +48,8 @@
         "img1": "cartuchera-unicornio-1.jpg",
         "img2": "cartuchera-unicornio-2.jpg",
         "img3": null,
-        "descripcion": null
+        "descripcion": null,
+        "categoria": "schoolSupplies"
     },
     {
         "id": 5,
@@ -57,7 +61,8 @@
         "img1": "color-filgo-x6.jpg",
         "img2": null,
         "img3": null,
-        "descripcion": null
+        "descripcion": null,
+        "categoria": "schoolSupplies"
     },
     {
         "id": 6,
@@ -69,7 +74,8 @@
         "img1": "color-filgo-x12.jpg",
         "img2": null,
         "img3": null,
-        "descripcion": null
+        "descripcion": null,
+        "categoria": "schoolSupplies"
     },
     {
         "id": 7,
@@ -81,7 +87,8 @@
         "img1": "lapiz-negro-x4.jpg",
         "img2": null,
         "img3": null,
-        "descripcion": null
+        "descripcion": null,
+        "categoria": "schoolSupplies"
     },
     {
         "id": 8,
@@ -93,7 +100,8 @@
         "img1": "lapiz-negro-x12.jpg",
         "img2": null,
         "img3": null,
-        "descripcion": null
+        "descripcion": null,
+        "categoria": "schoolSupplies"
     },
     {
         "id": 9,
@@ -105,7 +113,8 @@
         "img1": "goma.jpg",
         "img2": null,
         "img3": null,
-        "descripcion": null
+        "descripcion": null,
+        "categoria": "schoolSupplies"
     },
     {
         "id": 10,
@@ -117,7 +126,8 @@
         "img1": "sacapuntas.jpg",
         "img2": null,
         "img3": null,
-        "descripcion": null
+        "descripcion": null,
+        "categoria": "schoolSupplies"
     },
     {
         "id": 11,
@@ -129,7 +139,8 @@
         "img1": "lapicera-dino.jpg",
         "img2": "lapicera-unicornio.jpg",
         "img3": null,
-        "descripcion": null
+        "descripcion": null,
+        "categoria": "schoolSupplies"
     },
     {
         "id": 12,
@@ -141,7 +152,8 @@
         "img1": "lapicera-azul.jpg",
         "img2": "lapicera-negra.jpg",
         "img3": null,
-        "descripcion": null
+        "descripcion": null,
+        "categoria": "schoolSupplies"
     },
     {
         "id": 13,
@@ -153,7 +165,8 @@
         "img1": "lapicera-borrable.jpg",
         "img2": null,
         "img3": null,
-        "descripcion": "esta es una lapicera que esta muy buena"
+        "descripcion": null,
+        "categoria": "schoolSupplies"
     },
     {
         "id": 14,
@@ -165,7 +178,8 @@
         "img1": "cuaderno-a4-sin-anillos.jpg",
         "img2": null,
         "img3": null,
-        "descripcion": null
+        "descripcion": null,
+        "categoria": "schoolSupplies"
     },
     {
         "id": 15,
@@ -177,7 +191,8 @@
         "img1": "cuaderno-a4-anillado.jpg",
         "img2": null,
         "img3": null,
-        "descripcion": null
+        "descripcion": null,
+        "categoria": "schoolSupplies"
     },
     {
         "id": 16,
@@ -189,7 +204,8 @@
         "img1": "faber-x12.jpg",
         "img2": null,
         "img3": null,
-        "descripcion": null
+        "descripcion": null,
+        "categoria": "schoolSupplies"
     },
     {
         "id": 17,
@@ -201,6 +217,7 @@
         "img1": "faber-x15.jpg",
         "img2": null,
         "img3": null,
-        "descripcion": null
+        "descripcion": null,
+        "categoria": "schoolSupplies"
     }
 ]
