+++ views/index.ejs
@@ -24,7 +24,7 @@
             </article>
 
         <article class="schoolSupplies">
-            <a href="/productList#schoolSupplies"><h4 class="textCat">Productos escolares</h4></a>
+            <a href="/products#schoolSupplies"><h4 class="textCat">Productos escolares</h4></a>
             </article>
 
         <article class="planners">
