+++ views/partials/header.ejs
@@ -5,7 +5,7 @@
                 <a href="/">Inicio</a>
             </li>
             <li>
-                <a href="/productList">Productos</a>
+                <a href="/products">Productos</a>
             </li>
             <li>
                 <a href="/contact">Contacto</a>
