+++ views/products/productAdd.ejs
@@ -16,7 +16,7 @@
               <h1 class="title">Agregar Producto</h1>
 
               <div class="inputContainer">
-                  <input type="text" class="input" placeholder="a" required>
+                  <input type="text" class="input" placeholder="a" value="<%= productsList.length + 1 %>" disabled required>
                   <label for="" class="label">Id</label>
               </div>
