+++ views/products/productEdit.ejs
@@ -16,7 +16,7 @@
               <h1 class="title">Editar Producto</h1>
 
               <div class="inputContainer">
-                <input type="text" class="input" placeholder="a" required>
+                <input type="text" class="input" placeholder="a" value="<%=product.id%>" disabled required>
                 <label for="" class="label">Id</label>
             </div>
