+++ views/products/products.ejs
@@ -13,7 +13,7 @@
     <%- include('../partials/header') %> 
 
     <main>
-
+        
         <h2>Productos escolares</h2>
         <section id="schoolSupplies">
             
@@ -21,7 +21,7 @@
                 
             <article>
                 <h4><%= productsList[i].producto %></h4>
-                <a href="/productDetail/<%= productsList[i].id %> "><img src="./images/products/<%= productsList[i].img1 %>"></a>
+                <a href="/products/<%= productsList[i].id %> "><img src="./images/products/<%= productsList[i].img1 %>"></a>
                 
             </article>
